chore(code): Refactor for extensibility
